requirements.txt
python-doctr
python-doctr[tf]
opencv-python-headless==4.5.5.64
pandas
